Improve code readability and tests (#32)
